LPS-62070 Move StagingConfigurationPortletKeys and StagingProcessesPortletKeys to api
