+++ modules/apps/staging/staging-configuration-web/bnd.bnd
@@ -1,7 +1,6 @@
 Bundle-Name: Staging Configuration
 Bundle-SymbolicName: com.liferay.staging.configuration.web
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.staging.configuration.web.portlet.constants
 Include-Resource:\
 	classes,\
 	content=../staging-lang/src/main/resources/content,\
