+++ modules/apps/staging/staging-configuration-web/build.gradle
@@ -13,8 +13,8 @@ dependencies {
 	compile group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	compile group: "org.osgi", name: "org.osgi.compendium", version: "5.0.0"
 	compile project(":apps:application-list:application-list-api")
+	compile project(":apps:staging:staging-api")
 	compile project(":apps:staging:staging-lang")
-	compile project(":apps:staging:staging-processes-web")
 }
 
 liferay {
