+++ modules/apps/staging/staging-processes-web/build.gradle
@@ -15,6 +15,7 @@ dependencies {
 	compile project(":apps:application-list:application-list-api")
 	compile project(":apps:export-import:export-import-api")
 	compile project(":apps:layout:layout-taglib")
+	compile project(":apps:staging:staging-api")
 	compile project(":apps:staging:staging-lang")
 }
