+++ modules/apps/staging/staging-processes-web/src/main/java/com/liferay/staging/processes/web/portlet/action/EditPublishConfigurationMVCActionCommand.java
@@ -40,7 +40,7 @@ import com.liferay.portlet.exportimport.service.ExportImportConfigurationService
 import com.liferay.portlet.exportimport.staging.StagingUtil;
 import com.liferay.portlet.trash.service.TrashEntryService;
 import com.liferay.portlet.trash.util.TrashUtil;
-import com.liferay.staging.processes.web.constants.StagingProcessesPortletKeys;
+import com.liferay.staging.constants.StagingProcessesPortletKeys;
 
 import java.io.Serializable;
