+++ modules/apps/staging/staging-processes-web/src/main/java/com/liferay/staging/processes/web/portlet/action/ViewPublishConfigurationsMVCRenderCommand.java
@@ -15,7 +15,7 @@
 package com.liferay.staging.processes.web.portlet.action;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.staging.processes.web.constants.StagingProcessesPortletKeys;
+import com.liferay.staging.constants.StagingProcessesPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
