+++ modules/apps/staging/staging-processes-web/src/main/resources/META-INF/resources/init.jsp
@@ -90,8 +90,8 @@ page import="com.liferay.portlet.exportimport.service.ExportImportConfigurationL
 page import="com.liferay.portlet.exportimport.staging.LayoutStagingUtil" %><%@
 page import="com.liferay.portlet.exportimport.staging.StagingUtil" %><%@
 page import="com.liferay.portlet.trash.util.TrashUtil" %><%@
-page import="com.liferay.staging.configuration.web.portlet.constants.StagingConfigurationPortletKeys" %><%@
-page import="com.liferay.staging.processes.web.constants.StagingProcessesPortletKeys" %><%@
+page import="com.liferay.staging.constants.StagingConfigurationPortletKeys" %><%@
+page import="com.liferay.staging.constants.StagingProcessesPortletKeys" %><%@
 page import="com.liferay.staging.processes.web.search.PublishConfigurationDisplayTerms" %><%@
 page import="com.liferay.staging.processes.web.search.PublishConfigurationSearchTerms" %><%@
 page import="com.liferay.taglib.search.ResultRow" %><%@
